+++ src/extension.ts
@@ -1,48 +1,48 @@
 import * as vscode from "vscode";
 import { GDBDebugSession } from "./gdb";
 import { CoverageStatus } from './coverage';
+import { DebuggerSettings } from "./settings";
 
 const dockerTerminal = vscode.window.createTerminal("GnuCOBOL Docker");
-const dockerImage = 'olegkunitsyn/gnucobol:3.1-dev';
-const dockerMessage = 'GnuCOBOL Docker container is not defined in launch.json';
+const dockerMessage = "Property 'docker' is not defined in launch.json";
 
 export function activate(context: vscode.ExtensionContext) {
-    const containerStart = vscode.commands.registerCommand('gnucobol-debug.containerStart', function () {
+    const dockerStart = vscode.commands.registerCommand('gnucobol-debug.dockerStart', function () {
         const workspaceRoot: string = vscode.workspace.workspaceFolders[0].uri.fsPath;
         let config: vscode.DebugConfiguration;
         for (config of vscode.workspace.getConfiguration('launch', vscode.workspace.workspaceFolders[0].uri).get('configurations') as []) {
             if (config.type !== 'gdb') {
                 continue;
             }
-            if (config.container === undefined) {
+            if (config.docker === undefined) {
                 vscode.window.showInformationMessage(dockerMessage);
                 break;
             }
             dockerTerminal.show(true);
-            dockerTerminal.sendText(`docker run -d -i --name ${config.container} -w ${workspaceRoot} -v ${workspaceRoot}:${workspaceRoot} ${dockerImage}`);
+            dockerTerminal.sendText(`docker run -d -i --name gnucobol -w ${workspaceRoot} -v ${workspaceRoot}:${workspaceRoot} ${config.docker}`);
             break;
         };
     });
 
-    const containerStop = vscode.commands.registerCommand('gnucobol-debug.containerStop', function () {
+    const dockerStop = vscode.commands.registerCommand('gnucobol-debug.dockerStop', function () {
         let config: vscode.DebugConfiguration;
         for (config of vscode.workspace.getConfiguration('launch', vscode.workspace.workspaceFolders[0].uri).get('configurations') as []) {
             if (config.type !== 'gdb') {
                 continue;
             }
-            if (config.container === undefined) {
+            if (config.docker === undefined) {
                 vscode.window.showInformationMessage(dockerMessage);
                 break;
             }
             dockerTerminal.show(true);
-            dockerTerminal.sendText(`docker rm --force ${config.container}`);
+            dockerTerminal.sendText(`docker rm --force gnucobol`);
             break;
         };
     });
 
     context.subscriptions.push(
-        containerStart,
-        containerStop,
+        dockerStart,
+        dockerStop,
         vscode.debug.registerDebugConfigurationProvider('gdb', new GdbConfigurationProvider()),
         vscode.debug.registerDebugAdapterDescriptorFactory('gdb', new GdbAdapterDescriptorFactory(new CoverageStatus(), new GDBDebugSession())),
     );
@@ -55,11 +55,30 @@ export function deactivate() {
 class GdbConfigurationProvider implements vscode.DebugConfigurationProvider {
     resolveDebugConfiguration(folder: vscode.WorkspaceFolder | undefined, config: vscode.DebugConfiguration, token?: vscode.CancellationToken): vscode.ProviderResult<vscode.DebugConfiguration> {
         config.gdbargs = ["-q", "--interpreter=mi2"];
-        if (config.container !== undefined) {
+        if (config.docker !== undefined) {
             config.cobcpath = 'docker';
             config.gdbpath = 'docker';
-            config.cobcargs = ['exec', '-i', config.container, 'cobc'].concat(config.cobcargs);
-            config.gdbargs = ['exec', '-i', config.container, 'gdb'].concat(config.gdbargs);
+            config.cobcargs = ['exec', '-i', 'gnucobol', 'cobc'].concat(config.cobcargs);
+            config.gdbargs = ['exec', '-i', 'gnucobol', 'gdb'].concat(config.gdbargs);
+        }
+        const settings = new DebuggerSettings();
+        if (config.cwd === undefined) {
+            config.cwd = settings.cwd;
+        }
+        if (config.target === undefined) {
+            config.target = settings.target;
+        }
+        if (config.group === undefined) {
+            config.group = [];
+        }
+        if (config.arguments === undefined) {
+            config.arguments = "";
+        }
+        if (config.gdbpath === undefined) {
+            config.gdbpath = settings.gdbpath;
+        }
+        if (config.cobcpath === undefined) {
+            config.cobcpath = settings.cobcpath;
         }
         return config;
     }