+++ src/gdb.ts
@@ -30,7 +30,7 @@ class ExtendedVariable {
 export interface LaunchRequestArguments extends DebugProtocol.LaunchRequestArguments {
 	cwd: string;
 	target: string;
-	targetargs: string[];
+	arguments: string;
 	gdbpath: string;
 	gdbargs: string[];
 	cobcpath: string;
@@ -39,13 +39,13 @@ export interface LaunchRequestArguments extends DebugProtocol.LaunchRequestArgum
 	group: string[];
 	verbose: boolean;
 	coverage: boolean;
-	container: string;
+	docker: string;
 }
 
 export interface AttachRequestArguments extends DebugProtocol.LaunchRequestArguments {
 	cwd: string;
 	target: string;
-	targetargs: string[];
+	arguments: string;
 	gdbpath: string;
 	gdbargs: string[];
 	cobcpath: string;
@@ -69,12 +69,11 @@ export class GDBDebugSession extends DebugSession {
 	protected debugReady: boolean;
 	protected miDebugger: MI2;
 	coverageStatus: CoverageStatus;
-	private container: string;
+	private docker: string;
 	private showVariableDetails: boolean;
 	private settings = new DebuggerSettings();
 
 	protected initializeRequest(response: DebugProtocol.InitializeResponse, args: DebugProtocol.InitializeRequestArguments): void {
-		this.showVariableDetails = this.settings.displayVariableAttributes;
 		response.body.supportsSetVariable = true;
 		this.sendResponse(response);
 	}
@@ -83,7 +82,7 @@ export class GDBDebugSession extends DebugSession {
 		if (!args.coverage) {
 			this.coverageStatus = undefined;
 		}
-		this.container = args.container;
+		this.docker = args.docker;
 		this.started = false;
 		this.attached = false;
 
@@ -106,7 +105,7 @@ export class GDBDebugSession extends DebugSession {
 		this.crashed = false;
 		this.debugReady = false;
 		this.useVarObjects = false;
-		this.miDebugger.load(args.cwd, args.target, args.targetargs, args.group).then(() => {
+		this.miDebugger.load(args.cwd, args.target, args.arguments, args.group).then(() => {
 			setTimeout(() => {
 				this.miDebugger.emit("ui-break-done");
 			}, 50);
@@ -152,7 +151,7 @@ export class GDBDebugSession extends DebugSession {
 		this.crashed = false;
 		this.debugReady = false;
 		this.useVarObjects = false;
-		this.miDebugger.attach(args.cwd, args.target, args.targetargs, args.group).then(() => {
+		this.miDebugger.attach(args.cwd, args.target, args.arguments, args.group).then(() => {
 			setTimeout(() => {
 				this.miDebugger.emit("ui-break-done");
 			}, 50);
@@ -218,7 +217,7 @@ export class GDBDebugSession extends DebugSession {
 			return;
 
 		if (this.coverageStatus !== undefined) {
-			this.coverageStatus.show(this.miDebugger.getGcovFiles(), this.miDebugger.getSourceMap(), this.container);
+			this.coverageStatus.show(this.miDebugger.getGcovFiles(), this.miDebugger.getSourceMap(), this.docker);
 		}
 
 		this.quit = true;
@@ -405,6 +404,8 @@ export class GDBDebugSession extends DebugSession {
 	}
 
 	protected async variablesRequest(response: DebugProtocol.VariablesResponse, args: DebugProtocol.VariablesArguments): Promise<void> {
+		this.showVariableDetails = this.settings.displayVariableAttributes;
+		
 		let id: number | string | VariableObject | ExtendedVariable;
 		if (args.variablesReference < VAR_HANDLES_START) {
 			id = args.variablesReference - STACK_HANDLES_START;